requirements.txt
# Core Dependencies
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
feedparser>=6.0.10

# AI/ML Libraries
openai>=1.12.0
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
numpy>=1.24.0
scikit-learn>=1.3.0

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
pdf2image>=1.16.0
Pillow>=10.0.0

# Web Scraping
scrapy>=2.11.0
newspaper3k>=0.2.8
playwright>=1.40.0
selenium>=4.15.0

# HTML to PDF
pdfkit>=1.0.0
weasyprint>=60.0

# Data Processing
pandas>=2.0.0
python-dateutil>=2.8.0
pytz>=2023.3

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0

# Search
elasticsearch>=8.11.0
whoosh>=2.7.4

# Task Scheduling
APScheduler>=3.10.0
celery>=5.3.0
redis>=5.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
colorama>=0.4.6
tqdm>=4.66.0
urllib3>=2.0.0
robotexclusionrulesparser>=1.7.1

# Monitoring & Logging
loguru>=0.7.0
sentry-sdk>=1.39.0

# HTTP & APIs
urllib-robotparser>=1.0.0

# Command Line Interface
click>=8.1.7
rich>=13.7.0

# File handling
chardet>=5.2.0
python-magic>=0.4.27

# Image Processing (for figures)
opencv-python>=4.8.0

# NLP
spacy>=3.7.0
nltk>=3.8.0

# Validation
pydantic>=2.5.0
marshmallow>=3.20.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
